+++ KingBroker/Classes/HomePage(首页)/Controller/Report(报备)/WZSelectProjectsController.m
@@ -99,7 +99,7 @@ static NSString *size = @"20";
     NSMutableDictionary *paraments = [NSMutableDictionary dictionary];
     paraments[@"storeId"] = storeId;
     paraments[@"name"] = @"";
-    paraments[@"current"] = [NSString stringWithFormat:@"%zd",current];
+    paraments[@"current"] = [NSString stringWithFormat:@"%ld",(long)current];
     paraments[@"size"] = size;
     NSString *url = [NSString stringWithFormat:@"%@/projectCompany/projectList",URL];
     [mgr GET:url parameters:paraments progress:nil success:^(NSURLSessionDataTask * _Nonnull task, NSDictionary *  _Nullable responseObject) {
